# feat: early redirect for image