--- a/image/src/utils/constants.ts
+++ b/image/src/utils/constants.ts
@@ -1,7 +1,3 @@
-export const CACHE_DAY = 86400;
-export const CACHE_WEEK = CACHE_DAY * 7;
-export const CACHE_MONTH = CACHE_DAY * 30;
-
 export interface Env {
   MY_BUCKET: R2Bucket;
   DEDICATED_GATEWAY: string;
@@ -13,3 +9,16 @@ export interface Env {
   // wrangler secret
   IMAGE_API_TOKEN: string;
 }
+
+export const CACHE_SECOND = 1;
+export const CACHE_DAY = 86400;
+export const CACHE_WEEK = CACHE_DAY * 7;
+export const CACHE_MONTH = CACHE_DAY * 30;
+
+export const CACHE_TTL_BY_STATUS = {
+  cacheTtlByStatus: {
+    '200-299': CACHE_MONTH,
+    '404': CACHE_SECOND,
+    '500-599': 0,
+  },
+};