- [x] closes https://github.com/kodadot/nft-gallery/issues/4595

| before | after |
| --- | --- |
| <img width="1624" alt="Screenshot 2023-01-07 at 20 13 25" src="https://user-images.githubusercontent.com/734428/211153231-e310f7b2-ed01-4e9a-bfe4-47a211677f20.png"> | <img width="1624" alt="Screenshot 2023-01-07 at 20 12 23" src="https://user-images.githubusercontent.com/734428/211153252-f54a7767-6a0c-4e84-b511-e8f90980eebc.png"> |

on already uploaded images to cf-images and hit the fetch cache will be faster (<1s). on a new image still takes > 1s